Add packges duke and task task - contains the different task classes, duke contains the main logic
