+++ src/main/java/task/Todos.java
@@ -1,3 +1,5 @@
+package task;
+
 import java.io.FileWriter;
 import java.io.IOException;
